cleaned panel system file structure
